requirements.txt
mysqlclient
pyyaml
requests
sqlalchemy
ua-parser
user-agents